Implement support for Solis's authorization mechanism (see utilities/telegraf-plugins#9)
It returns the correct values for the examples given in their docs,
but I'm not currently able to test this against the API: I've not
been sent creds yet (in fact, my inverter's not even been installed yet)

It's a fairly simple signature mechanism, so there **shouldn't** be
much to go wrong with it - the most likely cause of issues will be
around the request body (the doc says that the hash should be empty
if the body is empty, but doesn't clarify whether they mean it
should be omitted, or a hash of a null string. The example in the
doc has a request body).