Implement support for rate-limit observance (utilities/telegraf-plugins#9)
This adds a `ratelimit` property to the object which is used to keep track
of the number of requests placed and the time that we started counting.

A rough overview is

- Every quota period (5s by default) the timer is reset.
- If there are more than `n` (3 by default) requests in the quota period, the request must wait

If a request is waiting, the quota will be rechecked every 1 second up to a maximum of `max_ratelimit_wait`
seconds (default 8). If the limit is hit, the script will exit with exit code `1`.

With the default config `max_ratelimit_wait` should never actually be hit and exists only as a safety net
against me writing in a bug which causes an infinite loop

It *might* be mildly interesting to have the script's LP output also provide details on its behaviour around
quotes (how often and how long it has waited etc)