Extract battery usage statistics and generate line protocol to expose the stats (utilities/telegraf-plugins#9)
This primarily uses the names exposed by the upstream API - once we have a more definitive idea  of what each
column exposes it may be worth giving more descriptive names.

This commit also adds a calculated field: `readingAge`

This field is an integer describing the time (in seconds) between the data being reported to the API
(i.e. `dataTimestamp`) and the reading being collected by the script.

It should allow us to build an alert to fire if the inverter stops reporting in