Add total consumption (utilities/telegraf-plugins#9)
This adds field `consumptionToday` to the inverter stats.

The figure represents total consumed, from all sources (i.e. solar, battery, grid etc).

It remains to be seen, however, whether it will suffer from the same issues as `gridPurchasedTodayEnergy` - it may
be that it too reports watts as whole numbers rather than a decimal.